+++ src/components/Landing/index.js
@@ -7,6 +7,7 @@ import "./landing.sass"
 import blob from "../../assets/blob.svg"
 import {withTranslation} from 'react-i18next'
 import Store from '../../db'
+import {AUTH} from '../../lib/firebase'
 
 
 class Landing extends Component {
@@ -20,6 +21,11 @@ class Landing extends Component {
 
   handleChange = ({target: {name, value}}) => this.setState({[name]: value})
 
+  handleSubmit = e => {
+    e.preventDefault && e.preventDefault()
+    this.handleUserLogin()
+  }
+
   handleUserLogin = () => {
     const {email, password} = this.state
     this.context.handleUserLogin(email, password)
@@ -27,74 +33,78 @@ class Landing extends Component {
 
   render() {
     const {t} = this.props
-    const {user} = this.context
 
     return (
-      <Centered>
-        {Object.keys(user).length ? <Redirect to={routes.DASHBOARD} /> : null}
-        <img alt={t("landing.blob-img-alt")} className="landing-blob landing-blob-1" src={blob}/>
-        <img alt={t("landing.blob-img-alt")} className="landing-blob landing-blob-2" src={blob}/>
-        <Typography variant="h4">{process.env.REACT_APP_TITLE}</Typography>
-        <Card style={{minWidth: "calc(100vw/3)", maxWidth: "calc(100vw - 32px)", margin: 16}}>
-          <CardContent>
-            <Grid container justify="center" spacing={8}>
-              <Grid item>
-                <TextField
-                  label={t("landing.email")}
-                  name="email"
-                  onChange={this.handleChange}
-                  type="email"
-                  variant="outlined"
-                />
-              </Grid>
-              <Grid item>
-                <TextField
-                  label={t("landing.password")}
-                  name="password"
-                  onChange={this.handleChange}
-                  type="password"
-                  variant="outlined"
-                />
-              </Grid>
-            </Grid>
-          </CardContent>
-          <CardActions>
-            <Grid alignItems="center" container justify="flex-end" spacing={16}>
-              <Grid item>
-                <MuiLink
-                  component={Link}
-                  style={{marginRight: 16}}
-                  to={routes.FORGOT_PASSWORD}
-                >{t("landing.forgot-password")}
-                </MuiLink>
-              </Grid>
-              <Grid container item justify="space-between" spacing={8}>
+      <form onSubmit={this.handleSubmit}>
+        <Centered>
+          {AUTH.currentUser ? <Redirect to={routes.DASHBOARD} /> : null}
+          <img alt={t("landing.blob-img-alt")} className="landing-blob landing-blob-1" src={blob}/>
+          <img alt={t("landing.blob-img-alt")} className="landing-blob landing-blob-2" src={blob}/>
+          <Typography variant="h4">{process.env.REACT_APP_TITLE}</Typography>
+          <Card style={{minWidth: "calc(100vw/3)", maxWidth: "calc(100vw - 32px)", margin: 16}}>
+            <CardContent>
+              <Grid container direction="column" justify="center" spacing={8}>
                 <Grid item>
-                  <Button
-                    color="primary"
-                    onClick={this.handleUserLogin}
-                    size="large"
-                    variant="contained"
-                  >
-                    {t("landing.login")}
-                  </Button>
+                  <TextField
+                    autoFocus
+                    fullWidth
+                    label={t("landing.email")}
+                    name="email"
+                    onChange={this.handleChange}
+                    type="email"
+                    variant="outlined"
+                  />
                 </Grid>
                 <Grid item>
-                  <Button
-                    color="secondary"
-                    component={Link}
-                    size="large"
-                    to={routes.REGISTER}
-                  >
-                    {t("landing.register")}
-                  </Button>
+                  <TextField
+                    fullWidth
+                    label={t("landing.password")}
+                    name="password"
+                    onChange={this.handleChange}
+                    type="password"
+                    variant="outlined"
+                  />
                 </Grid>
               </Grid>
-            </Grid>
-          </CardActions>
-        </Card>
-        <Typography variant="caption">{process.env.REACT_APP_VERSION}</Typography>
-      </Centered>
+            </CardContent>
+            <CardActions>
+              <Grid alignItems="center" container justify="flex-end" spacing={16}>
+                <Grid alignItems="center" container item justify="space-evenly" spacing={8}>
+                  <Grid item>
+                    <Button
+                      color="secondary"
+                      onClick={this.handleUserLogin}
+                      size="large"
+                      type="submit"
+                      variant="contained"
+                    >
+                      {t("landing.login")}
+                    </Button>
+                  </Grid>
+                  <Grid item>
+                    <MuiLink
+                      component={Link}
+                      style={{marginRight: 16}}
+                      to={routes.FORGOT_PASSWORD}
+                    >{t("landing.forgot-password")}
+                    </MuiLink>
+                  </Grid>
+                  <Grid item>
+                    <Button
+                      color="primary"
+                      component={Link}
+                      size="large"
+                      to={routes.REGISTER}
+                    >
+                      {t("landing.register")}
+                    </Button>
+                  </Grid>
+                </Grid>
+              </Grid>
+            </CardActions>
+          </Card>
+        </Centered>
+      </form>
     )
   }
 }