+++ src/components/shared/index.js
@@ -1,0 +1,13 @@
+import Loading from "./Loading"
+import DestructButton from "./DestructButton"
+import Dialog from "./Dialog"
+import Page, {withPage} from "./Page"
+
+
+export {
+  DestructButton,
+  Loading,
+  Page,
+  Dialog,
+  withPage
+}