+++ src/lib/material-ui.js
@@ -1,13 +1,7 @@
 import {createMuiTheme} from '@material-ui/core/styles'
 import deepmerge from 'deepmerge'
 
-const defaultTheme = createMuiTheme({typography: {useNextVariants: true}})
-
-
-const primaryFontFamily = {fontFamily: "Poppins, Arial, sans-serif"}
-
-const secondaryFontFamily = {fontFamily: "adobe-garamond-pro, 'Times New Roman'"}
-
+const secondaryFontFamily = "adobe-garamond-pro, 'Times New Roman'"
 
 const base = {
   palette: {
@@ -25,21 +19,10 @@ const base = {
   },
   typography: {
     useNextVariants: true,
-
-    h1: {
-      fontFamily: primaryFontFamily
-    },
-    h2: {
-      fontFamily: primaryFontFamily
-    },
-    h3: {
-      fontFamily: primaryFontFamily
-    },
-    button: {
-      fontFamily: primaryFontFamily
-    },
+    fontSize: 16,
+    fontFamily: "Poppins, Arial, sans-serif",
     h4: {
-      fontFamily: secondaryFontFamily
+      fontWeight: "bolder"
     },
     h5: {
       fontFamily: secondaryFontFamily
@@ -51,7 +34,7 @@ const base = {
       fontFamily: secondaryFontFamily
     },
     subtitle2: {
-      color: "#000000C0"
+      color: "#00000080"
     },
     body1: {
       fontFamily: secondaryFontFamily
@@ -70,8 +53,10 @@ const base = {
     MuiButton: {
       sizeLarge: {
         padding: '8px 24px',
-        fontSize: defaultTheme.typography.pxToRem(20),
         borderRadius: 24
+      },
+      containedPrimary: {
+        boxShadow: "none"
       },
       containedSecondary: {
         boxShadow: "none"
@@ -83,18 +68,23 @@ const base = {
 
 const darkTheme = {
   palette: {
-    type: "dark"
-  },
-  /*List:{
-    ListItem: {
-      color: "#fff"
+    type: "dark",
+    primary: {
+      dark: "#51c1b7",
+      main: "#98dad4",
+      light: "#f2f9f8"
+    },
+    secondary: {
+      dark: "#02405a",
+      main: "#6d8c9b",
+      light: "#ced9de",
+      contrastText: "#fff"
     }
-  },*/
+  },
   TextField: {
     color: "#fff"
   },
   typography: {
-    useNextVariants: true,
 
     h1: {
       color: "#fff"
@@ -141,17 +131,18 @@ const darkTheme = {
     },
     MuiFormControl:{
       root:{
-        backgroundColor: "#f2f9f8"
+        backgroundColor: "#333",
+        color: "#000"
       }
     },
-    MuiSwitch: {
-      colorPrimary: {
-        '&$checked': {
-          color: "#51c1b7",
-          '& + $bar': {
-            backgroundColor: "#51c1b7"
-          }
-        }
+    MuiBottomNavigation: {
+      root: {
+        backgroundColor: "#333333"
+      }
+    },
+    MuiAvatar: {
+      colorDefault: {
+        color: "#fff"
       }
     }
   }